move `assert_sql_matches` into asserts module
